step ➜
## Overview
This "journal" is an experiment into capturing thoughts, usage, etc. related to step-wise commits along with the commit. Much of the content should be updated to match each commit. Some of the content will persist across some number of commits.
## Usage
There's no usage to add in this commit. This "usage" section is added as a placeholder.
These notes should be cleaned up and provided as content in the [README](../README.md).
- To enable pretty syntax highlighting in editors, `.envrc` files and the like (`.env` too) should start with the `#!/bin/env bash` header.